backend/pyproject.toml
[project]
name = "jlo-backend"
version = "1.0.0"
description = "jlo - Just Logs: A simple logging server"
requires-python = ">=3.8"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.20",
    "pydantic>=2.10.0",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.3.0",
    "pygrok>=1.0.0",
    "python-dateutil>=2.9.0.post0",
    "argon2-cffi>=25.1.0",
    "schedule>=1.2.0",
]

[dependency-groups]
dev = []
